backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
aiomysql==0.2.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Milvus
pymilvus==2.3.4

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Text Processing
langchain==0.1.6
langchain-text-splitters==0.0.1

# Ollama
httpx==0.26.0

# DeepSeek
openai==1.12.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2